refactor(eslint): simplify ESLint config by removing extends array and manually including rules
The ESLint configuration was refactored to remove the `extends` array and manually include the recommended rules from `js` and `tseslint`. This change improves maintainability by making the configuration more explicit and easier to modify in the future. Additionally, the `DateRange` type import was consolidated in the date-picker component to reduce redundancy.